+++ spec/unit/optimizer/test_extend.rb
@@ -2,16 +2,97 @@ require 'spec_helper'
 module Bmg
   describe "extend optimization" do
 
-    context "extend.restrict" do
-      let(:relation) {
-        Relation.new([
-          { a: 1,  b: 2 },
-          { a: 11, b: 2 }
-        ])
+    let(:relation) {
+      Relation.new([
+        { a: 1,  b: 2 },
+        { a: 11, b: 2 }
+      ])
+    }
+
+    let(:c_ext) {
+      ->(t){ 12 }
+    }
+
+    let(:d_ext) {
+      ->(t){ 13 }
+    }
+
+    context 'extend with empty extension' do
+      subject {
+        relation.extend({})
       }
 
+      it 'returns the extension itself' do
+        expect(subject).to be(relation)
+      end
+    end
+
+    context "extend.allbut" do
+      subject {
+        relation.extend(extension).allbut(butlist)
+      }
+
+      context 'when the butlist covers the whole extension' do
+        let(:extension) {
+          { c: c_ext, d: d_ext }
+        }
+        let(:butlist){
+          [:a, :c, :d]
+        }
+
+        # rel(:a, :b).extend(:c, :d).allbut(:a, :c, :d) => [:b]
+        # becomes
+        # rel(:a, :b).allbut(:a)
+        it 'strips the extension completely and simplifies the butlist' do
+          expect(subject).to be_a(Operator::Allbut)
+          expect(subject.send(:butlist)).to eql([:a])
+          expect(operand(subject)).to be(relation)
+        end
+      end
+
+      context 'when the butlist touches the extension, but not whole' do
+        let(:extension) {
+          { c: c_ext, d: d_ext }
+        }
+        let(:butlist){
+          [:a, :c]
+        }
+
+        # rel(:a, :b).extend(:c, :d).allbut(:a, :c) => [:b, :d]
+        # becomes
+        # rel(:a, :b).extend(:d).allbut(:a) => [:b, :d]
+        it 'simplifies the extension' do
+          expect(subject).to be_a(Operator::Allbut)
+          expect(subject.send(:butlist)).to eql([:a])
+          expect(operand(subject)).to be_a(Operator::Extend)
+          expect(operand(subject).send(:extension)).to eql(d: d_ext)
+        end
+      end
+
+      context 'when the butlist does not touch the extension' do
+        let(:extension) {
+          { c: c_ext, d: d_ext }
+        }
+        let(:butlist){
+          [:a, :b]
+        }
+
+        # rel(:a, :b).extend(:c, :d).allbut(:a, :b) => [:c, :d]
+        # becomes
+        # rel(:a, :b).extend(:c, :d).allbut(:a, :c) => [:c, :d]
+        it 'does not optimize at all' do
+          expect(subject).to be_a(Operator::Allbut)
+          expect(subject.send(:butlist)).to eql(butlist)
+          expect(operand(subject)).to be_a(Operator::Extend)
+          expect(operand(subject).send(:extension)).to eql(extension)
+        end
+      end
+    end
+
+    context "extend.restrict" do
+
       let(:extension) {
-        { c: ->(t){ 12 } }
+        { c: c_ext }
       }
 
       subject{
@@ -44,12 +125,6 @@ module Bmg
     end
 
     context "extend.page" do
-      let(:relation) {
-        Relation.new([
-          { a: 1,  b: 2 },
-          { a: 11, b: 2 }
-        ])
-      }
 
       subject {
         relation.extend(extension).page([:a], 1, page_size: 2)
@@ -57,7 +132,7 @@ module Bmg
 
       context 'when the ordering does not touch the extension' do
         let(:extension) {
-          { c: ->(t){ 12 } }
+          { c: c_ext }
         }
 
         it 'pushes the page down' do
@@ -76,6 +151,69 @@ module Bmg
 
         it 'does not optimize' do
           expect(subject).to be_a(Operator::Page)
+        end
+      end
+    end
+
+    context "extend.project" do
+
+      subject {
+        relation.extend(extension).project(projection)
+      }
+
+      # rel(:a, :b).extend(:c).project(:a) => [:a]
+      # becomes
+      # rel(:a, :b).project(:a) => [:a]
+      context 'when the projection does not touch the extension at all' do
+        let(:extension) {
+          { c: c_ext }
+        }
+        let(:projection){
+          [:a]
+        }
+
+        it 'strips the extension completely' do
+          expect(subject).to be_a(Operator::Project)
+          expect(subject.send(:attrlist)).to eql(projection)
+          expect(operand(subject)).to be(relation)
+        end
+      end
+
+      # rel(:a, :b).extend(:c, :d).project(:a, :c, :d) => [:a, :c, :d]
+      # becomes
+      # rel(:a, :b).extend(:c, :d).project(:a, :c, :d) => [:a, :c, :d]
+      context 'when the projection covers the whole extension' do
+        let(:extension) {
+          { c: c_ext, d: d_ext }
+        }
+        let(:projection){
+          [:a, :c, :d]
+        }
+
+        it 'does not optimize at all' do
+          expect(subject).to be_a(Operator::Project)
+          expect(subject.send(:attrlist)).to eql(projection)
+          expect(operand(subject)).to be_a(Operator::Extend)
+          expect(operand(subject).send(:extension)).to eql(extension)
+        end
+      end
+
+      # rel(:a, :b).extend(:c, :d).project(:a, :c) => [:a, :c]
+      # becomes
+      # rel(:a, :b).extend(:c).project(:a, :c) => [:a, :c]
+      context 'when the projection covers part of the extension' do
+        let(:extension) {
+          { c: c_ext, d: d_ext }
+        }
+        let(:projection){
+          [:a, :c]
+        }
+
+        it 'simplifies them' do
+          expect(subject).to be_a(Operator::Project)
+          expect(subject.send(:attrlist)).to eql(projection)
+          expect(operand(subject)).to be_a(Operator::Extend)
+          expect(operand(subject).send(:extension)).to eql(c: c_ext)
         end
       end
     end